Add utility to execute an operation with redraw disabled for a control
The method Control#setRedraw(boolean) usually has to be executed as a
pair of disablement and succeeding re-enablement. In case the statements
executed between disablement and re-enabled may throw an exception, this
has to be handled in every such execution sequence to avoid a missing
re-enablement upon an occurring exception.

This change introduces a utility method that encapsulates the execution
of statements while disabling redraw for a defined control. This
implementation ensures the restoration of the redraw state even if the
executed statements throw an exception.