Set proper environment for Maven build depending on Windows architecture
Maven builds accessing the binaries (i.e., executing tests) currently
fail on Windows when in aarch64 is used. This is because the x86
binaries are accessed as both architectures are defined in the
environment setup for Windows. With this change, the proper environment
is set depending on current architecture (x86/aarch).